moving plan_alchem_networksutils
