+++ openfecli/commands/plan_rhfe_network.py
@@ -16,7 +16,7 @@ from typing import List
 from openfecli.utils import write
 from openfecli import OFECommandPlugin
 from openfecli.parameters import MOL_DIR, MAPPER, OUTPUT_DIR
-from openfecli.commands.plan_alchemical_networks_utils import plan_alchemical_network_output
+from openfecli.plan_alchemical_networks_utils import plan_alchemical_network_output
 
 
 def plan_rhfe_network_main(
